Account and Transaction Group most of the work is done.

+++ PersonalExpenseManagerModel/src/com/pratikabu/pem/model/utils/impl/SearchFacadeImpl.java
@@ -106,11 +106,16 @@ public class SearchFacadeImpl implements SearchFacade {
 		Session session = HibernateConfiguration.getFactory().getCurrentSession();
 		session.beginTransaction();//to let know hibernate that the object below exists we have begin the transaction
 		try {
-			for(Object t : toBeDeleted) {
-				session.delete(t);//it does not delete it here
+			if(null != toBeDeleted) {
+				for(Object t : toBeDeleted) {
+					session.delete(t);//it does not delete it here
+				}
 			}
-			for(Object t : toBeSaved) {
-				session.saveOrUpdate(t);//it does not save it here
+			
+			if(null != toBeSaved) {
+				for(Object t : toBeSaved) {
+					session.saveOrUpdate(t);//it does not save it here
+				}
 			}
 			session.getTransaction().commit();//it saves the object here
